+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/preview-middleware
 
+## 0.20.20
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.13
+
 ## 0.20.19
 
 ### Patch Changes