+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.1.13
+
+### Patch Changes
+
+-   Updated dependencies [15ec5c4]
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.9
+    -   @sap-ux/inquirer-common@0.7.6
+    -   @sap-ux/odata-service-inquirer@2.4.12
+
 ## 0.1.12
 
 ### Patch Changes